Remove old ArchiveBase
